% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amu}
\alias{amu}
\title{Antimicrobial usage (AMU)}
\format{A data frame with 3664 rows and 2 variables:
\describe{
  \item{WeekID}{number of the week, integer.}
  \item{AMU}{usage of antimicrobials, logical.}
}}
\source{
\url{http://viparc.org}
}
\usage{
amu
}
\description{
Data collected in the ViParc project (http://viparc.org).
}
\details{
Generated from \code{amu <- read.csv2(system.file("extdata", "AMU_week.csv", package = "viparc"), sep=",", na.strings = "na", header = TRUE, stringsAsFactors = FALSE)}
}
\keyword{datasets}
