% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amu_causes_disease.R
\name{amu_causes_disease}
\alias{amu_causes_disease}
\title{Effect of AMU on occurence of disease}
\usage{
amu_causes_disease(x, time = "WEEK", space = "FlockID", amu = "AMU",
  disease = "Sick_yes", week_amu = 1, week_disease = 2)
}
\arguments{
\item{x}{a data frame.}

\item{time}{the name of the data frame variable that corresponds to time (the
data frame will be orderd chronologically).}

\item{week_amu}{the number of weeks before the focal week we consider to
define the presence or absence of AMU.}

\item{week_disease}{the number of weeks after the focal week over which we
look for presence or absence of disease.}

\item{time}{the name of the data frame variable that corresponds to space (
the computation of the presence of AMU and disease will be
performed by spatial unit).}

\item{time}{the name of the data frame variable that corresponds to AMU
(logical).}

\item{time}{the name of the data frame variable that corresponds to presence
of disease (logical).}
}
\value{
A list containing the contingency table and the result of the Fisher
        exact test.
}
\description{
Uses a Fisher exact test to test whether AMU increases disease occurence.
}
\details{
We consider only weeks with no diseases. Among these weeks, we make 2 groups:
one with AMU and one without AMU. The groups with AMU contains the weeks that
have AMU. The groups without AMU contains the weeks that do not have AMU as
well as no AMU of the \code{week_amu} weeks that precede. For these 2 groups,
we look at the presence of disease during the \code{week_disease} weeks that
follow. We thus end up with a contingency table of AMU and presence of
disease that we feed to a Fisher exact test.
}
\examples{
library(dplyr)

amu_with_disease \%>\%
  amu_causes_diseases()

amu_with_disease \%>\%
  amu_causes_diseases(disease = "DIARRHOEA")

amu_with_disease \%>\%
  mutate(Sick_yes = DIARRHOEA | RESPIRATORY ) \%>\%
  amu_causes_diseases()

amu_with_disease \%>\%
  mutate(Sick_yes = DIARRHOEA | RESPIRATORY ) \%>\%
  amu_causes_diseases(week_disease = 3)

}
