% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amu_with_diseases}
\alias{amu_with_diseases}
\title{AMU and Diseases data}
\format{A data frame with 3663 rows and 18 variables:
\describe{
  \item{WeekID}{ID of week, combining ID of flock with ID of week.}
  \item{AMU}{usage of antimicrobials, logical.}
  \item{FarmID}{ID of farm.}
  \item{FlockID}{ID of flock, combining ID of farm with ID of cycle.}
  \item{Cycle}{ID of cycle.}
  \item{WEEK}{Week number.}
  \item{RESPIRATORY}{Presence / absence of respiratory symptoms.}
  \item{DIARRHOEA}{Presence / absence of diarrhoea symptoms.}
  \item{CNS}{Presence / absence of central nervous system infection symptoms.}
  \item{MALAISE}{Presence / absence of malaise.}
  \item{LEGLESIONS}{Presence / absence of leg lesions.}
  \item{SUDDENDEATH}{Presence / absence of sudden death.}
  \item{NoSign}{Presence / absence of  symptoms.}
  \item{Sick_yes}{Presence / absence of symptoms (any symptom).}
  \item{OTHDISEASE}{Presence / absence of other disease.}
  \item{CHICKENSICKNO}{Number of sick chicken.}
  \item{CHICKENDISEASEDEATH}{Number of chicken dead, excluding sudden death.}
  \item{CHICKENSUDDENDEATH}{Number of chicken dying from sudden death.}
}}
\source{
\url{http://viparc.org}
}
\usage{
amu_with_diseases
}
\description{
Data collected in the ViParc project (http://viparc.org).
}
\details{
Generated from \code{amu_with_diseases <- merge(AMU, diseases, by = c("WeekID"), all.X = TRUE)}
}
\keyword{datasets}
