% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_disease.R
\name{any_disease}
\alias{any_disease}
\title{Look for presence of disease}
\usage{
any_disease(x, n = 1)
}
\arguments{
\item{x}{A boolean vector informing about the presence or absence of disease
at a given point in time. This implies that this vector is
considered to be chronologically ordered, the most recent
observations being first.}

\item{n}{A integer scalar telling the number of time points after the current
one to look for for defining the period of time over which we want
to look for presence of disease.}
}
\value{
A boolean of vector of the same length as x, the n final values of
        which are NA.
}
\description{
Look for the presence of disease over a period that follows a given time
point.
}
\examples{
disease <- rep(rep(c(FALSE, TRUE), 7),
               c(rep(1, 4), 2, 3, 2, 2, 1, 1, 3, 2, 2, 3))
cbind(disease, any_disease(disease, 3))

}
