% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chcknfarms}
\alias{chcknfarms}
\title{Disease and antimicrobial surveillance in chicken farms}
\format{A data frame with 2226 rows and 9 variables:
\describe{
  \item{FARMCODE}{code of the farm, character.}
  \item{CYCLE}{number of the cycle, integer.}
  \item{WEEK}{number of the week, integer}
  \item{SRepiratory}{presence of respiratory infection, logical.}
  \item{SDiarrhoea}{presence of diarrhoea, logical.}
  \item{SCNS}{presence of central nervous system infection, logical.}
  \item{SMalaise}{presence of malaise, logical.}
  \item{Week_Mt}{weekly mortality, numeric.}
  \item{AntibioticUse}{use of antibiotic, logical.}
}}
\source{
\url{http://viparc.org}
}
\usage{
chcknfarms
}
\description{
Data collected in the ViParc project (http://viparc.org).
}
\keyword{datasets}
