% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_gaps.R
\name{fill_gaps}
\alias{fill_gaps}
\title{Fill gaps}
\usage{
fill_gaps(x, n = 1)
}
\arguments{
\item{x}{A vector of boolean.}

\item{n}{A scalar integer specifying the length of gaps that we want to fill.}
}
\value{
A vector of boolean of the same size as \code{x}.
}
\description{
This function fills the gaps in a sequence of booleans.
}
\details{
If \code{n = 1}, it will replace any \code{TRUE FALSE TRUE} sequence by
\code{TRUE TRUE TRUE}. If \code{n = 1}, it will additionally replace any
\code{TRUE FALSE FALSE TRUE} sequence by \code{TRUE TRUE TRUE TRUE} and so on.
}
\examples{
x <- rep(c(rep(c(FALSE, TRUE), 7), FALSE),
         c(1, 3, 1, 1, 1, 2, 1, 2, 3, 2, 3, 1, 2, 3, 1))
fill_gaps(x)
fill_gaps(x, 2)

}
