% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_map.R
\name{heat_map}
\alias{heat_map}
\title{Heatmap of disease and antimicrobial usage of surveillance data}
\usage{
heat_map(df, dec = FALSE, xlab = "time", ylab = "space",
  col_na = "lightgrey", col_surv_no_dis = "blue", col_surv_dis = "red",
  ...)
}
\arguments{
\item{df}{data frame with 4 variable named "time", "space", "disease" and
"drug". "time" and "space" are considered as integer and character
respectively and "disease" and "drug" are both considered as
logical.}

\item{xlab}{label of the x axis. Default to "time".}

\item{ylab}{label of the y axis. Default to "space".}

\item{col_na}{color used for non surveyed time/space.}

\item{col_surv_no_dis}{color used for surveyed time/space with no disease.}

\item{col_surv_dis}{color used for surveyed time/space with disease}

\item{...}{additional parameters passed to the "points" function that add
the "drug" information.}
}
\description{
blablabla
}
\examples{
library(dplyr)
chcknfarms \%>\%
  filter(CYCLE == 1) \%>\%
  transmute(space   = FARMCODE,
            time    = WEEK,
            disease = SRespiratory,
            drug    = AntibioticUse) \%>\%
  heat_map(xlab = "week", ylab = "farm")

}
\author{
Marc Choisy
}
