% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_map.R
\name{order_rows}
\alias{order_rows}
\title{order the rows of a matrix according to the column index of the last non null value.}
\usage{
order_rows(mat, dec)
}
\description{
order the rows of a matrix according to the column index of the last non null value.
}
