% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_index.R
\name{time_index}
\alias{time_index}
\title{Give the index of a sequence}
\usage{
time_index(x)
}
\arguments{
\item{x}{a vector.}
}
\value{
A vector of integers of the same length as \code{x}.
}
\description{
From a sequence \code{A A A B B C A A C C C} it will returns \code{1 2 3 1 2
1 1 2 1 2 3}.
}
\details{
This is basically a wrapper around the \code{\link[base]{rle}} function.
}
\examples{
x <- c("A", "A", "A", "B", "B", "C", "A", "A", "C", "C", "C")
time_index(x)

}
